Version 1.1.4
Better C++11 conformance.
Test with QT 5.12.3.

+++ googledrive/browse.cpp
@@ -218,7 +218,7 @@ CBrowseReply CPluginObject::browse (QString const & objectID, CContentDirectory:
           }
         }
 
-        int numberReturned = items.size ();
+        unsigned numberReturned = static_cast<unsigned>(items.size ());
         reply.setNumberReturned (numberReturned);
         reply.setTotalMatches (numberReturned);
       }
